arch/stm32{|f0l0g0|f7|h5|h7|l4}/stm32_adc.c: add support for batch DMA transfer
Add an option that configure the number of regular group conversions
that will triger a DMA callback transfering data to the upper-half driver.

By default this value is 1 and the driver behaves the same as before
the change.

DMA support for H5 and H7 is not complete so this change has no effect,
but for consistency they have also been modified.

Signed-off-by: raiden00pl <raiden00@railab.me>